third_party/blink/web_tests/fast/dom/Range/collapsed-range-bounding-client-rect-expected.txt

This is a testharness.js-based test.
[FAIL] Check position for collapsed range.
  assert_equals: Position of collapsed rect should match right edge of non-collapsed range. expected "16, 273.515625" but got "16, 273.5"
Harness: the test ran to completion.
